etl/requirements.txt
celery==5.4.0
redis==5.0.3
schedule==1.2.1
requests==2.31.0
pandas==2.2.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
python-dotenv==1.0.1
